+++ src/app/login-guard.ts
@@ -7,7 +7,7 @@ import { State } from './reducers';
 import { AuthService } from './services/auth.service';
 
 @Injectable()
-export class Guard implements CanActivate {
+export class LoginGuard implements CanActivate {
 
   constructor(private authService: AuthService, private store: Store<State>) {}
